% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\docType{package}
\name{mdl-package}
\alias{mdl}
\alias{mdl-package}
\title{mdl: Performant drop-in replacement for model matrices}
\description{
Performant drop-in replacement for model matrices.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/simonpcouch/mdl}
  \item Report bugs at \url{https://github.com/simonpcouch/mdl/issues}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
