% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_setup.R
\name{mdl_api_open_documentation}
\alias{mdl_api_open_documentation}
\title{Open MDL APIs documentation in browser}
\usage{
mdl_api_open_documentation()
}
\value{
Does not return any value
}
\description{
Prints the url to the APIs documentation webpage and opens it in the browser.
}
\examples{
# Simply call the function to see the documentation
mdl_api_open_documentation()
}
