% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_setup.R
\name{mdl_api_set_key}
\alias{mdl_api_set_key}
\title{Set API key}
\usage{
mdl_api_set_key(api_key)
}
\arguments{
\item{api_key}{The API key.}
}
\description{
Sets the API key. The key can be found in your MDL profile after you log-in (https://microdata.unhcr.org/index.php/auth/profile).
}
