% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_setup.R
\name{mdl_api_set_url}
\alias{mdl_api_set_url}
\title{Set API URL}
\usage{
mdl_api_set_url(enum_api_url)
}
\arguments{
\item{enum_api_url}{The API base endpoint URL. We recommend to use the related enumerator, e.g.: mdl_enum_api_url$uat}
}
\description{
Sets the API base endpoint URL. You can specify any of the different versions on the platform (production or user testing) and call the function again for switching from one to another.
}
