% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collection.R
\name{mdl_collection_get}
\alias{mdl_collection_get}
\title{Get one collection}
\usage{
mdl_collection_get(collection_id)
}
\arguments{
\item{collection_id}{The ID of the collection. To see an up to date list of collections call mdl_collection_list(). You can also use the corresponding enumerator, e.g.: mdl_enum_collection$WestAfrica}
}
\value{
API call response.
}
\description{
Returns a single collection
}
