% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collection.R
\name{mdl_datasets_by_collection}
\alias{mdl_datasets_by_collection}
\title{List datasets by collections}
\usage{
mdl_datasets_by_collection(limit = 500, offset = 0)
}
\arguments{
\item{limit}{Max number of datasets to get.}

\item{offset}{Point from which the query will start, to be used if you want to used paging. Default is 0.}
}
\value{
API call response.
}
\description{
List all datasets with collection owner and all linked collections.
There is one row for each linked collection included the main collection, meaning that datasets repeat.
}
