% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external_wbg.R
\name{mdl_harvest_nada}
\alias{mdl_harvest_nada}
\title{Harvest from external NADA}
\usage{
mdl_harvest_nada(
  library_url = "https://microdata.worldbank.org",
  survey_idno,
  base_path = NULL,
  upload_files = FALSE,
  idno_prefix = "WBG",
  enum_collection = mdl_enum_collection$WorldBank,
  overwrite,
  published
)
}
\arguments{
\item{library_url}{The url of the library, for example: "https://microdata.worldbank.org"}

\item{survey_idno}{Unique survey identifier in NADA}

\item{base_path}{The path on your computer where the files will be downloaded, if not specified uses the current working directory}

\item{upload_files}{TRUE if you want to upload a copy of the documentation files; FALSE if you want just to put a direct link to the files}

\item{idno_prefix}{A string to be used as prefix of the idno, so that it can be easily recognized as harvested.}

\item{enum_collection}{The collection where copy the dataset. Please consider using mdl_enum_collection.}

\item{overwrite}{TRUE if you want to overwrite the dataset in case it already exists; FALSE if you want to avoid to overwrite}

\item{published}{TRUE if you want to publish directly the dataset; FALSE if you want just to load it in the back-end}
}
\value{
API call response.
}
\description{
Copies a dataset from another NADA microdata library to the UNHCR MDL.
By default, harvests from the WBG library (https://microdata.worldbank.org). Changing the parameters, the harvest should work with other NADA based libraries (given other implementations and versions it may not work though).
}
\details{
Warning messages about incomplete final line found in readLines is not a concern. Please note that files get downloaded only if not present in the base_path folder: to have up to date files delete them.
}
