% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource.R
\name{mdl_resource_file_delete}
\alias{mdl_resource_file_delete}
\title{Delete a resource}
\usage{
mdl_resource_file_delete(survey_idno, file_id)
}
\arguments{
\item{survey_idno}{The unique identifier of an existing survey, e.g.: UNHCR_ETH_2020_SENS_v2.1}

\item{resource_id}{Resource ID. Use mdl_resource_list() to see the IDs

TODO: this does not work, it deletes the whole survey!!!}
}
\value{
API call response.
}
\description{
Deletes a resource from a survey.
}
