% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource.R
\name{mdl_resource_upload_file_and_link_old}
\alias{mdl_resource_upload_file_and_link_old}
\title{Upload a file and create a resource linked to it}
\usage{
mdl_resource_upload_file_and_link_old(
  survey_idno,
  file_path,
  enum_resource_type,
  resource_title,
  resource_description = NULL
)
}
\arguments{
\item{survey_idno}{The unique identifier of an existing survey, e.g.: UNHCR_ETH_2020_SENS_v2.1}

\item{file_path}{Path to the file to upload, e.g.: C:/path/to/file.pdf}

\item{enum_resource_type}{The type of resource. We recommend to use the corresponding enumerator, e.g.: mdl_enum_resource_type$doc_report}

\item{resource_title}{Title of the resource as shown in the platform to the final user.}

\item{resource_description}{Brief description of the resource as shown in the platform to the final user.}
}
\value{
API call response.
}
\description{
This function first uploads a file for a given survey, then creates a resource and link it to the uploaded file.
}
