% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey.R
\name{mdl_survey_attach_to_collections}
\alias{mdl_survey_attach_to_collections}
\title{Attach studies to collections}
\usage{
mdl_survey_attach_to_collections(
  survey_idno,
  enum_collection = NULL,
  link_collections,
  mode = "update"
)
}
\arguments{
\item{survey_idno}{Dataset IDNo.}

\item{enum_collection}{Main collection of the dataset (optional). It is recommended to use mdl_enum_collection.}

\item{link_collections}{Array containing other secondary collections in which the dataset has to be shown (required). It is recommended to use mdl_enum_collection.}

\item{mode}{Select flag to update or replace existing linked collections for the study: "replace" = replace linked collections for the study with the provided list; "update" = (Default) add/update linked collections list.}
}
\value{
API call response.
}
\description{
Used to modify the main collection and/or add the secondary collections.
In case you only want to change the main collection it is recommended to use mdl_survey_options(), or you can try to set link_collections to the same value of enum_collection.
}
