% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey.R
\name{mdl_survey_delete}
\alias{mdl_survey_delete}
\title{Delete a survey given an idno}
\usage{
mdl_survey_delete(survey_idno)
}
\arguments{
\item{survey_idno}{Survey unique identifier}
}
\value{
API call response.
}
\description{
Given an unique idno, deletes the survey
}
