% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey.R
\name{mdl_survey_download_ddi}
\alias{mdl_survey_download_ddi}
\title{Download survey DDI file}
\usage{
mdl_survey_download_ddi(survey_idno, path = NULL)
}
\arguments{
\item{survey_idno}{Path to the DDI XML file.}

\item{path}{Path to where you want to save the file, if not specified uses current working directory.}
}
\value{
File path
}
\description{
Downloads a DDI XML file.
}
