% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_pdf.R
\name{mdl_survey_generate_pdf}
\alias{mdl_survey_generate_pdf}
\title{Create documentation PDF}
\usage{
mdl_survey_generate_pdf(
  survey_idno,
  variable_list = 0,
  variable_description = 0,
  external_resource = 0
)
}
\arguments{
\item{survey_idno}{The identifier of the survey for which you want to generate the PDF, e.g.: UNHCR_ETH_2020_SENS_v2.1}

\item{variable_list}{Specifies if the PDF should include the variable list: 1 for yes, 0 for no.}

\item{variable_description}{Specifies if the PDF should include the variable description: 1 for yes, 0 for no.}

\item{external_resource}{Specifies if the PDF should include the external resources: 1 for yes, 0 for no.}
}
\value{
API call response.
}
\description{
This function creates the documentation PDF from the survey metadata and links a resource to it.
}
