% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{mdl_survey_get_variable}
\alias{mdl_survey_get_variable}
\title{Get a variable}
\usage{
mdl_survey_get_variable(survey_idno, variable_id)
}
\arguments{
\item{survey_idno}{Survey unique identifier}

\item{variable_id}{Variable ID}
}
\value{
API call response.
}
\description{
Fetches the metadata for a specific variable like numeric stats etc.
If only need name and label you can use mdl_survey_get_variables.
}
