% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{mdl_survey_get_variable_files}
\alias{mdl_survey_get_variable_files}
\title{Get variables data files}
\usage{
mdl_survey_get_variable_files(survey_idno)
}
\arguments{
\item{survey_idno}{Survey unique identifier}
}
\value{
API call response.
}
\description{
Get data files metadata containing variables.
Can be used when getting all variables to filter them by file.
}
