% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{mdl_survey_get_variables}
\alias{mdl_survey_get_variables}
\title{Get all variables}
\usage{
mdl_survey_get_variables(survey_idno, data_file = NA)
}
\arguments{
\item{survey_idno}{Survey unique identifier}

\item{data_file}{Data file ID to be provided to get variables for only one specific data file (if not provided all variables will returned). Can be found using the function mdl_survey_get_variable_files() or it can be seen in the "fid" column in the result of this function itself.}
}
\value{
API call response.
}
\description{
Fetches list of variables for a given dataset, this contains variable name and label.
It can be filtered using optional argument data file to get variables for only one data file (you can filter after receiving the unfiltered result as well since there is a column with the data file ID).
}
