% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey.R
\name{mdl_survey_import_ddi}
\alias{mdl_survey_import_ddi}
\title{Create survey from DDI file}
\usage{
mdl_survey_import_ddi(
  xml_file,
  rdf_file = NULL,
  enum_collection,
  enum_survey_access_policy,
  data_remote_url = NULL,
  published = FALSE,
  overwrite = FALSE
)
}
\arguments{
\item{xml_file}{Path to the DDI XML file.}

\item{rdf_file}{Path to the RDF file.}

\item{enum_collection}{The ID of the collection where the survey will be created. To see an up to date list of collections call mdl_collection_list().You can also use the corresponding enumerator, e.g.: mdl_enum_collection$WestAfrica}

\item{enum_survey_access_policy}{Specifies the access level to the data files. You can use the corresponding enumerator, e.g.: mdl_enum_survey_access_policy$'Licensed use files'}

\item{data_remote_url}{Link to the data files, in case enum_survey_access_policy is set to remote (link to external repository).}

\item{published}{The survey status: FALSE for draft, TRUE for published.}

\item{overwrite}{Specifies if the survey will be overwritten in case it already exists: FALSE for not overwriting, TRUE for overwriting. If a survey with the same idno already exists and the argument is set to FALSE, the survey will not change and an error will be returned.}
}
\value{
API call response
}
\description{
Imports a DDI XML file and its metadata creating a new survey.
}
