% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey.R
\name{mdl_survey_internal_id}
\alias{mdl_survey_internal_id}
\title{Get survey internal ID}
\usage{
mdl_survey_internal_id(survey_idno)
}
\arguments{
\item{survey_idno}{Survey unique identifier}
}
\value{
Dataset internal ID.
}
\description{
Given an unique idno, it returns the internal id.
}
