% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey.R
\name{mdl_survey_list}
\alias{mdl_survey_list}
\title{Get survey list}
\usage{
mdl_survey_list(limit = 50, offset = 0)
}
\arguments{
\item{limit}{Max number of datasets to get. Default is 50.}

\item{offset}{Point from which the query will start, to be used if you want to used paging. Default is 0.}
}
\value{
API call response.
}
\description{
Gets list of surveys with main info
}
