% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey.R
\name{mdl_survey_url}
\alias{mdl_survey_url}
\title{Get link to survey}
\usage{
mdl_survey_url(survey_idno)
}
\arguments{
\item{survey_idno}{Survey unique identifier}
}
\value{
Url to dataset.
}
\description{
Given an unique idno, it return the link to that survey. Please note that this will be redirected to another url format that uses an internal id instead of the idno.
}
