% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtrx.R
\name{mtrx}
\alias{mtrx}
\title{Modern model matrices}
\usage{
mtrx(formula, data)
}
\arguments{
\item{formula}{A formula. Cannot contain inlined functions like \code{-}or \code{*}.}

\item{data}{A data frame.}
}
\description{
\code{mdl::mtrx()} (read: "model matrix") implements an opinionated and performant
reimagining of model matrices. It takes in a formula and data frame,
like \code{model.frame()}, and outputs a numeric matrix of predictor values,
like \code{model.matrix()}.
}
\section{Comparison to Base R}{


Compared to \code{model.matrix()}, \code{mdl::mtrx()}:
\itemize{
\item Does not accept formulae with inlined functions (like \code{-} or \code{*}).
\item Never drops rows (and thus doesn't accept an \code{na.action}).
\item Assumes that factors levels are encoded as they're intended (i.e. \code{drop.unused.levels} and \code{xlev} are not accepted).
}

\code{mdl::mtrx()} is intended to be paired with the recipes package for
preprocessing.
}

\examples{
mdl::mtrx(mpg ~ ., mtcars)

}
