% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{survey_get_variable}
\alias{survey_get_variable}
\title{Get a a variable}
\usage{
survey_get_variable(survey_idno, variable_id)
}
\arguments{
\item{survey_idno}{Survey unique identifier}

\item{variable_id}{VAriable ID}
}
\value{
API call response.
}
\description{
Fetches the metadata for a variable
}
