% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl-package.R
\docType{package}
\name{mdl-package}
\alias{mdl}
\alias{mdl-package}
\title{mdl: Modern model matrices}
\description{
An opinionated and performant reimagining of model matrices using Rust.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/simonpcouch/mdl}
  \item \url{https://simonpcouch.github.io/mdl/}
  \item Report bugs at \url{https://github.com/simonpcouch/mdl/issues}
}

}
\author{
\strong{Maintainer}: Simon Couch \email{simon.couch@posit.co} (\href{https://orcid.org/0000-0001-5676-5107}{ORCID})

Other contributors:
\itemize{
  \item Mossa M. Reimert \email{mossa@sund.ku.dk} [contributor]
  \item Posit Software, PBC [copyright holder, funder]
}

}
\keyword{internal}
