% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource.R
\name{mdl_resource_file_list}
\alias{mdl_resource_file_list}
\title{Get list of resources}
\usage{
mdl_resource_file_list(survey_idno)
}
\arguments{
\item{survey_idno}{The unique identifier of an existing survey, e.g.: UNHCR_ETH_2020_SENS_v2.1}
}
\value{
API call response.
}
\description{
Given a survey identifier, provides list of resources
}
